Don't try to validate Linux CIS profile compliance on Windows
