+++ pkg/cli/cmds/profile_windows.go
@@ -1,0 +1,11 @@
+//go:build windows
+// +build windows
+
+package cmds
+
+import (
+	"github.com/urfave/cli"
+)
+
+func validateProfile(clx *cli.Context, role CLIRole) {
+}
